+++ x/params/CHANGELOG.md
@@ -25,4 +25,16 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ## [Unreleased]
 
+<<<<<<< HEAD:x/params/CHANGELOG.md
 ## [v0.2.0-rc.1](https://github.com/cosmos/cosmos-sdk/releases/tag/x/params/v0.2.0-rc.1) - 2024-12-18
+=======
+> Testing suite for the core module. Should stay in 0ver. 
+
+## [v0.0.2](https://github.com/cosmos/cosmos-sdk/releases/tag/core/testing%2Fv0.0.2)
+
+* [#23396](https://github.com/cosmos/cosmos-sdk/pull/23396) Add branch service mock.
+
+## [v0.0.1](https://github.com/cosmos/cosmos-sdk/releases/tag/core/testing%2Fv0.0.1)
+
+* Initial tag.
+>>>>>>> 9ea03a3c4 (chore: remove baseapp and `sdk.Context` fully from x/upgrade (#23396)):core/testing/CHANGELOG.md